% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Lab Logger UI}
\usage{
app_ui(app_title = "Lab Logger", app_color = "red", timezone = NULL)
}
\arguments{
\item{app_title}{the title of the application}

\item{app_color}{the dashboard color, see \link[shinydashboard]{dashboardPage} skin for available options}
}
\description{
Generates the user interface part of the isoviewer app
}
