% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{c3_add_device}
\alias{c3_add_device}
\title{Add new device}
\usage{
c3_add_device(device_name, desc = NA, device_type_id = "undefined",
  group_id = default(group_id), particle_id = NA, in_use = TRUE,
  con = default(con), quiet = default(quiet))
}
\arguments{
\item{device_name}{device name / id, must be unique for each group (will error if not)}

\item{desc}{device description}

\item{device_type_id}{device type name, must exist in database (will error if not)}

\item{group_id}{group name, must exist in database (will error if not)}

\item{particle_id}{optional, will be automatically filled in the first time the device logs to the database}

\item{in_use}{whether device is in use (if not, cannot log any data)}
}
\description{
Add new device
}
