% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{ll_add_experiment_devices}
\alias{ll_add_experiment_devices}
\title{Link device to experiment}
\usage{
ll_add_experiment_devices(exp_id, device_names, data_idxs,
  device_ids = ll_get_device_ids(device_names, quiet = quiet),
  data_key_prefix = NA, group_id = default(group_id), con = default(con),
  quiet = default(quiet))
}
\arguments{
\item{exp_id}{the unique identifier of the experiment (usually a few letter code). Has to exist already in the data base.}

\item{device_names}{names of the device(s) to link (define alternatively to the device id, will determine device_ids internally)}

\item{data_idxs}{the data indices to map for the experiment, must be same length as device_names/device_ids}

\item{device_ids}{one or multiple device ids to add (by default is inferred from the device names)}
}
\description{
Note that for the same device, experiment and index, only one record can exist (will error if already exists).
}
