% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{c3_add_group}
\alias{c3_add_group}
\title{Add a new group record}
\usage{
c3_add_group(group_id, desc = NA, con = default(con),
  quiet = default(quiet))
}
\arguments{
\item{group_id}{the unique identifier/name of the group. Will error if id already exists.}

\item{desc}{the description for the device group (optional)}
}
\description{
Add a new group record
}
