% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{c3_calculate_duration}
\alias{c3_calculate_duration}
\title{Convert datetime to duration}
\usage{
c3_calculate_duration(df, units)
}
\arguments{
\item{df}{data frame with datetime column}

\item{units}{any time unit that lubridate understands}
}
\value{
df with new column \code{duration}
}
\description{
Does not need to be sorted upfront, makes different to minimum datetime within the group_by groupings.
}
