% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{c3_get_device_ids}
\alias{c3_get_device_ids}
\title{Get device id}
\usage{
c3_get_device_ids(device_names, group_id = default(group_id),
  quiet = default(quiet))
}
\description{
Convenience shortcut function to get the database integer device identifier for one or multiple devices.
}
