% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{c3_get_experiment_devices}
\alias{c3_get_experiment_devices}
\title{Retrieve experiment devices}
\usage{
c3_get_experiment_devices(group_id = default(group_id), filter = NULL,
  select = c(exp_id, exp_name, recording, device_id, device_name, particle_id,
  data_key_prefix, data_idx), con = default(con), quiet = default(quiet))
}
\arguments{
\item{group_id}{experiments from which group to fetch}
}
\value{
experiments_devices
}
\description{
Returns experiment-device links (only for active/in-use devices) joined with experiment and devices tables so filter conditions can be applied on the devices as well.
}
