% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{c3_get_experiments}
\alias{c3_get_experiments}
\title{Retrieve experiments}
\usage{
c3_get_experiments(group_id = default(group_id), filter = NULL,
  con = default(con), quiet = default(quiet))
}
\arguments{
\item{group_id}{experiments from which group to fetch}
}
\value{
experiments
}
\description{
Retrieve experiments
}
