% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{c3_info_messages}
\alias{c3_info_messages}
\alias{c3_turn_info_messages_on}
\title{Turn information messages on/off}
\usage{
c3_turn_info_messages_on(data = NULL)
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}
}
\description{
These functions turn information messages on/off in all subsequent function calls by changing the global settings for the \code{quiet} parameter of c3 functions.
These functions can be called stand alone or within a pipeline to turn messages on/off at a certain point during the pipeline.
}
