% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{c3_plot_device_data_logs}
\alias{c3_plot_device_data_logs}
\title{Plot device data logs}
\usage{
c3_plot_device_data_logs(device_data_logs, filter = NULL,
  duration_units = NULL, date_breaks = NULL, quiet = default(quiet))
}
\arguments{
\item{duration_units}{specify a time unit (e.g. "mins") to indicate whether x axis should be displayed as a duration since the first data point within each experiment, if NULL x axis is displayed as regular date time.}

\item{date_breaks}{formate the datetime breaks if not plotting duration (i.e. is ignored if duration_units is provided)}
}
\description{
Plot device data logs
}
