% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{c3_set_access_token}
\alias{c3_set_access_token}
\title{Set default access token for particle API requests}
\usage{
c3_set_access_token(token)
}
\arguments{
\item{token}{the particle acocunt access token, keep secret!!}
}
\description{
Set default access token for particle API requests
}
