% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_cloud.R
\name{c3_test_experiment_devices_data}
\alias{c3_test_experiment_devices_data}
\title{Test which values one gets for a set of experiment devices}
\usage{
c3_test_experiment_devices_data(experiment_devices, spread = FALSE,
  access_token = default(access_token), quiet = default(quiet))
}
\arguments{
\item{experiment_devices}{experiment_devices records, see \link{c3_get_experiment_devices}}
}
\description{
Test which values one gets for a set of experiment devices
}
