% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_device_info.R
\name{deviceInfoServer}
\alias{deviceInfoServer}
\title{Device Info Server}
\usage{
deviceInfoServer(
  input,
  output,
  session,
  get_cloud_state,
  refresh_cloud_state,
  get_cloud_data,
  refresh_cloud_data,
  get_cloud_info,
  refresh_cloud_info,
  get_device_ids,
  get_state_logs,
  refresh_state_logs
)
}
\description{
Device Info Server
}
