% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{generate_gs_token}
\alias{generate_gs_token}
\title{authenticate for google spreadsheets and store token in target directory (allows GUI to access the spreadsheet data)}
\usage{
generate_gs_token(data_dir)
}
\arguments{
\item{data_dir}{the data directory where the GUI is intended to run}
}
\description{
authenticate for google spreadsheets and store token in target directory (allows GUI to access the spreadsheet data)
}

