% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lablogger-package}
\alias{lablogger}
\alias{lablogger-package}
\title{lablogger: Lab Logger}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/lablogger}
  \item Report bugs at \url{https://github.com/kopflab/lablogger/issues}
}

}
\keyword{internal}
