% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{labwareC3-package}
\alias{labwareC3}
\alias{labwareC3-package}
\title{labwareC3: Chemostat Control Center}
\description{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/labwareC3}
  \item Report bugs at \url{https://github.com/kopflab/labwareC3/issues}
}

}
\keyword{internal}
