% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{ll_add_device_type}
\alias{ll_add_device_type}
\title{Add new device type}
\usage{
ll_add_device_type(
  device_type_id,
  device_type_desc,
  con = default(con),
  quiet = default(quiet)
)
}
\arguments{
\item{device_type_id}{device type ID, must be unique (will error if not)}

\item{device_type_desc}{device type description}
}
\description{
Add new device type
}
