% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_get_device_data_logs}
\alias{ll_get_device_data_logs}
\title{Retrieve data logs}
\usage{
ll_get_device_data_logs(group_id = default(group_id), filter = NULL,
  select = c(exp_id, device_id, device_name, device_data_log_id, datetime,
  data_idx, starts_with("data_key"), starts_with("data_")), max_rows = NULL,
  convert_to_TZ = Sys.getenv("TZ"), con = default(con),
  quiet = default(quiet))
}
\arguments{
\item{filter}{what filter conditions to apply, if any (forwarded to \link[dplyr]{filter})}

\item{select}{what columns to select (forwarded to \link[select]{select}), by default a selection of the most commonly used columns}

\item{max_rows}{if provided, only selects the indicated number of rows (more efficient this way than part of the filter)}

\item{convert_to_TZ}{if provided, converts the log_datetime to the provided timezone (by default the local one stored in \code{Sys.getenv("TZ")}). If NULL, will keep it as UTC.}
}
\value{
device state logs
}
\description{
Returns data logs joined with experiments, devices, and experiment_device_data (for prefix only) tables so filter conditions can be applied on these as well. Sorted by descending order (i.e. latest record first).
}
