% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_get_device_state_logs}
\alias{ll_get_device_state_logs}
\title{Retrieve state logs}
\usage{
ll_get_device_state_logs(
  group_id = default(group_id),
  filter = NULL,
  select = c(device_id, device_name, device_state_log_id, log_datetime, log_type,
    log_message, starts_with("state"), notes),
  max_rows = NULL,
  convert_to_TZ = Sys.timezone(),
  con = default(con),
  quiet = default(quiet)
)
}
\arguments{
\item{filter}{what filter conditions to apply, if any (forwarded to \link[dplyr]{filter})}

\item{select}{what columns to select (forwarded to \link[select]{select}), by default a selection of the most commonly used columns}

\item{max_rows}{if provided, only selects the indicated number of rows (more efficient this way than part of the filter)}

\item{convert_to_TZ}{if provided, converts the log_datetime to the provided timezone (by default the local one stored in \code{Sys.timezone()}). If NULL, will keep it as UTC.}
}
\value{
device state logs
}
\description{
Returns state logs joined with devices table so filter conditions can be applied on the devices as well. Sorted by descending order (i.e. latest record first).
}
