% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_get_device_types}
\alias{ll_get_device_types}
\title{Retrieve device types}
\usage{
ll_get_device_types(
  select = everything(),
  con = default(con),
  quiet = default(quiet)
)
}
\value{
device types
}
\description{
Retrieve device types
}
