% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_get_devices}
\alias{ll_get_devices}
\title{Retrieve devices}
\usage{
ll_get_devices(
  group_id = default(group_id),
  filter = NULL,
  select = c(device_id, particle_id, device_name, device_type_id, device_type_desc),
  in_use_only = TRUE,
  con = default(con),
  quiet = default(quiet)
)
}
\arguments{
\item{group_id}{devices from which group to fetch}
}
\value{
devices
}
\description{
Retrieve devices
}
