% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_cloud.R
\name{ll_get_devices_cloud_info}
\alias{ll_get_devices_cloud_info}
\title{Get general device information}
\usage{
ll_get_devices_cloud_info(
  devices = ll_get_devices(group_id = group_id, con = con),
  include_unregistered = FALSE,
  group_id = default(group_id),
  con = default(con),
  access_token = default(access_token),
  convert_to_TZ = Sys.timezone(),
  quiet = default(quiet)
)
}
\arguments{
\item{devices}{data frame with devices for which to get cloud info, by default all devices associated with the group}

\item{include_unregistered}{whether to provide cloud info for devices that are not registered in the database (no by default)}

\item{access_token}{the access token for the accout}

\item{particle_id}{the ID(s) of the particle device(s)}
}
\value{
nested data frame (converted from JSON)
}
\description{
Get information from the particle cloud about devices.
}
