% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_cloud.R
\name{ll_get_devices_cloud_state}
\alias{ll_get_devices_cloud_state}
\title{Get device state}
\usage{
ll_get_devices_cloud_state(devices = ll_get_devices(group_id = group_id,
  con = con), group_id = default(group_id), con = default(con),
  access_token = default(access_token), convert_to_TZ = Sys.timezone(),
  spread = FALSE, quiet = default(quiet))
}
\arguments{
\item{access_token}{the access token for the accout}

\item{convert_to_TZ}{if provided, converts the log_datetime to the provided timezone (by default the local one stored in \code{Sys.getenv("TZ")}). If NULL, will keep it as UTC.}

\item{spread}{whether to convert the state data into wide format (note that this combines value and units columns!)}
}
\value{
nested data frame (converted from JSON)
}
\description{
Get state from the particle cloud for devices.
}
