% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_get_exp_device_data_logs}
\alias{ll_get_exp_device_data_logs}
\title{Get device data logs for specific experiment(s)}
\usage{
ll_get_exp_device_data_logs(exp_id, group_id = default(group_id), ...,
  cache = TRUE, read_cache = TRUE, quiet = default(quiet))
}
\arguments{
\item{exp_id}{experiment ID(s)}

\item{...}{forwarded to ll_get_device_state_logs}

\item{cache}{whether to cache the data logs}

\item{reac_cache}{whether to read the cache}
}
\description{
Also supports efficient caching of the retrieved data.
}
