% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_get_experiment_device_links}
\alias{ll_get_experiment_device_links}
\title{Retrieve experiment device links}
\usage{
ll_get_experiment_device_links(group_id = default(group_id),
  filter = NULL, select = c(exp_device_data_id, exp_id, recording,
  device_id, device_name, particle_id, data_group, data_idx, active),
  con = default(con), quiet = default(quiet))
}
\arguments{
\item{group_id}{experiments from which group to fetch}
}
\value{
experiments_devices
}
\description{
Returns experiment-device links (only for active/in-use devices) joined with experiment, devices and device types tables so filter conditions can be applied on any of these as well.
}
