% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_get_experiments}
\alias{ll_get_experiments}
\title{Retrieve experiments}
\usage{
ll_get_experiments(
  group_id = default(group_id),
  filter = NULL,
  convert_to_TZ = Sys.timezone(),
  con = default(con),
  quiet = default(quiet)
)
}
\arguments{
\item{group_id}{experiments from which group to fetch}
}
\value{
experiments
}
\description{
Retrieve experiments
}
