% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ll_plot_device_data_logs}
\alias{ll_plot_device_data_logs}
\title{Plot device data logs}
\usage{
ll_plot_device_data_logs(device_data_logs, filter = NULL,
  show_error_range = FALSE, exclude_outliers = FALSE,
  duration_breaks = NULL, date_breaks = NULL,
  include_device_info = FALSE, overlay_experiments = FALSE,
  quiet = default(quiet))
}
\arguments{
\item{device_data_logs}{data logs}

\item{duration_breaks}{specify a duration unit or interval (e.g. "hours" or "20 mins") to indicate whether x axis should be displayed as a duration since the first data point within each experiment. If interval is specified, this will be used for axis intervals, if just time unit the axis will have regular auto-ticks. If duration_breaks = NULL (the default), the x axis is displayed as regular date time (set date_breaks instead).}

\item{date_breaks}{formate the datetime breaks if not plotting duration (i.e. is ignored if duration_units is provided)}

\item{include_device_info}{whether to include the device info (name and index) in the data trace label}

\item{overlay_experiments}{whether to overlay the experiments instead of paneling (default is panels). This usually makes most sense if x axis is a duration (set via duration units)}
}
\description{
This function helps to visualize data logs retrieved with \code{\link{ll_get_device_data_logs}} or read in from downloaded data logs.
}
\seealso{
Other data logs functions: \code{\link{ll_get_device_data_logs}},
  \code{\link{ll_read_device_data_logs_from_file}},
  \code{\link{ll_write_device_data_logs_to_file}}
}
\concept{data logs functions}
