% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read.R
\name{ll_read_device_data_logs_from_file}
\alias{ll_read_device_data_logs_from_file}
\title{Read data logs}
\usage{
ll_read_device_data_logs_from_file(file_path, quiet = default(quiet))
}
\arguments{
\item{file_path}{path to the .rds or .xlsx file}
}
\value{
device data logs data frame
}
\description{
Read in an .rds, .csv or .xlsx file with device data logs downloaded from the GUI.
}
\seealso{
Other data logs functions: \code{\link{ll_get_device_data_logs}},
  \code{\link{ll_plot_device_data_logs}},
  \code{\link{ll_write_device_data_logs_to_file}}
}
\concept{data logs functions}
