% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{ll_remove_experiment_device_links}
\alias{ll_remove_experiment_device_links}
\title{Remove experiment device links}
\usage{
ll_remove_experiment_device_links(
  exp_id,
  exp_device_data_ids,
  con = default(con),
  quiet = default(quiet)
)
}
\arguments{
\item{exp_id}{experiment id}

\item{exp_device_data_ids}{exp_device_data_ids}
}
\description{
Removes/deactivates the specified experiment device links. Removes those that don't already have data records associated with them and deactivates the others.
}
