% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_reset_exp_device_data_logs_cache}
\alias{ll_reset_exp_device_data_logs_cache}
\title{Reset device data logs cache}
\usage{
ll_reset_exp_device_data_logs_cache(exp_id, quiet = default(quiet))
}
\arguments{
\item{exp_id}{experiment ID(s)}
}
\description{
Reset device data logs cache
}
