% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{ll_run_gui}
\alias{ll_run_gui}
\title{Run the user interface}
\usage{
ll_run_gui(group_id, access_token, pool, timezone = Sys.timezone(),
  app_pwd = NULL, app_title = group_id, app_color = "red", ...,
  launch = FALSE)
}
\arguments{
\item{group_id}{which group to run for}

\item{access_token}{access token for particle account}

\item{pool}{ideally database connection pool, see \link[pool]{dbPool} but can also be a single db connection (not recommended)}

\item{app_title}{the title of the application}

\item{app_color}{the dashboard color, see \link[shinydashboard]{dashboardPage} skin for available options}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call (only if \code{launch = TRUE}), can include server-specific parameters such as host or port}

\item{launch}{whether to launch the app (TRUE) or return a shiny app object (FALSE) that then can be launched via \code{\link[shiny]{runApp}}
(note: if \code{launch=FALSE}, \code{...} gets ignored)}

\item{password}{which password to require for login. If NULL, login will be automatic. (NOTE: maybe manage by data base at some point?)}
}
\description{
This function runs the user interface for the lab logger.
}
