% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_cloud.R
\name{ll_send_devices_command}
\alias{ll_send_devices_command}
\title{Send device commands to the cloud}
\usage{
ll_send_devices_command(devices, command, message = "",
  access_token = default(access_token), quiet = default(quiet))
}
\arguments{
\item{devices}{data frame with devices for which to issue commands}

\item{command}{to send}

\item{message}{message to add to command}

\item{access_token}{the access token for the accout}
}
\value{
nested data frame (converted from JSON)
}
\description{
Send commands to one or more devices.
}
