% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{ll_set_group_id}
\alias{ll_set_group_id}
\title{Set default group ID}
\usage{
ll_set_group_id(group_id)
}
\arguments{
\item{group_id}{group}
}
\description{
Note that this is not checked wheter it exists, simply used as the default for other functions.
}
