% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{ll_set_request_timeout}
\alias{ll_set_request_timeout}
\title{Set default request timeout}
\usage{
ll_set_request_timeout(timeout)
}
\arguments{
\item{group_id}{group}
}
\description{
Set default request timeout
}
