% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_setup.R
\name{ll_setup_gui}
\alias{ll_setup_gui}
\title{Setup the user interface app}
\usage{
ll_setup_gui(dir, overwrite = FALSE)
}
\arguments{
\item{dir}{the directory where to create the launch scripts (absolute path or relative to your current working directory)}

\item{overwrite}{whether to overwrite existing launch files (default is FALSE)}
}
\description{
Sets up the necessary files for launching the user interface in the target folder. Keeps sensitive authentication information separate from the app launch script and allows for easy running of the gui via \code{\link[shiny]{runApp}} or deployment as a server application (e.g. on shinyapps.io). IMPORTANT: never add the created \code{credentials.R} file to any publically accessible shared folder or github repository - the database and particle account information you have stored there could be compromised!
}
