% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_snyc_device_names_from_cloud}
\alias{ll_snyc_device_names_from_cloud}
\title{Synchronize device names}
\usage{
ll_snyc_device_names_from_cloud(
  in_use_only = TRUE,
  con = default(con),
  access_token = default(access_token),
  quiet = default(quiet)
)
}
\value{
devices with updated device names
}
\description{
Synchronize device names
}
\note{
no longer necessary, sync happens during logging
Update device names in the database from the cloud.
}
