% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{ll_summarize_data_logs}
\alias{ll_summarize_data_logs}
\title{Generate data log summary}
\usage{
ll_summarize_data_logs(
  data_logs,
  exclude_outliers = FALSE,
  slope_denom_units = "hour"
)
}
\description{
Generate data log summary
}
