% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{ll_update_experiment_info}
\alias{ll_update_experiment_info}
\title{Update experiment information}
\usage{
ll_update_experiment_info(exp_id, exp_desc = NULL, exp_notes = NULL,
  group_id = default(group_id), con = default(con),
  quiet = default(quiet))
}
\arguments{
\item{exp_id}{the unique identifier of the experiment (usually a few letter code). Will error if id already exists.}

\item{group_id}{group name, must exist in database (will error if not)}
}
\value{
whether the update was successful or not
}
\description{
Update experiment information
}
