% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_write.R
\name{ll_write_device_state_logs_to_file}
\alias{ll_write_device_state_logs_to_file}
\title{Save device state logs}
\usage{
ll_write_device_state_logs_to_file(device_state_logs, ...)
}
\description{
Save device state logs
}
