% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_plot_download.R
\name{plotDownloadLink}
\alias{plotDownloadLink}
\title{Plot Download Link}
\usage{
plotDownloadLink(id, label = "Save")
}
\arguments{
\item{label}{Label for the download link}
}
\description{
Plot Download Link
}
