% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{run}
\alias{run}
\title{Start the user interface}
\usage{
run(db_host, db_name, db_user, db_pwd, db_drv = PostgreSQL, ...,
  launch = TRUE)
}
\arguments{
\item{db_host}{data base host}

\item{db_name}{data base name}

\item{db_user}{data base user}

\item{db_pwd}{data base user password}

\item{db_drv}{data base driver function, default is PostgreSQL}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call (only if \code{launch = TRUE}), can include server-specific parameters such as host or port}

\item{launch}{whether to launch the app (TRUE) or return a shiny app object (FALSE) that then can be launched via \code{\link[shiny]{runApp}}
(note: if \code{launch=TRUE}, \code{...} gets ignored)}
}
\description{
Run the user interface for the chemostat control center
}
