% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{run_gui}
\alias{run_gui}
\title{Start the user interface}
\usage{
run_gui(base_dir = ".", ...)
}
\arguments{
\item{base_dir}{the directory for local data files}

\item{...}{passed on to the \link{shiny::runApp} call and could include parameters such as host or port}
}
\description{
To run this function stand-alone as a script from any folder on your system, copy the
run (Unix systems) or run.bat (Windows) systems from the GitHub repository
(\link{http://www.github.com/KopfLab/labwareC3/inst/scripts/}
and modify depending on your R executable.
}

