% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{run_gui_dev}
\alias{run_gui_dev}
\title{for GUI development
not exported to namespace but helpful for active development of the shiny app}
\usage{
run_gui_dev(base_dir = ".", app_dir = file.path("inst", "gui"), ...)
}
\description{
for GUI development
not exported to namespace but helpful for active development of the shiny app
}

