% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{update_package}
\alias{update_package}
\title{install the latest version of labwareC3 from GitHub}
\usage{
update_package(ref = "master")
}
\arguments{
\item{ref}{which version to install, master (=newest) is the default}
}
\description{
install the latest version of labwareC3 from GitHub
}
