% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Chemostat Control Center Server}
\usage{
app_server(group_id, access_token, pool, app_pwd, timezone,
  start_screen = "data")
}
\description{
Generates the server part of the isoviewer app
}
