% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{c3_set_db_con}
\alias{c3_set_db_con}
\title{Set default database connection or pool}
\usage{
c3_set_db_con(con)
}
\arguments{
\item{con}{database connection or pool object}
}
\description{
Set default database connection or pool
}
