% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_device_info.R
\name{deviceInfoServer}
\alias{deviceInfoServer}
\title{Device Info Server}
\usage{
deviceInfoServer(input, output, session, get_cloud_state,
  refresh_cloud_state, get_cloud_data, refresh_cloud_data,
  refresh_experiment_device_links, get_cloud_info, refresh_cloud_info)
}
\arguments{
\item{refresh_experiment_device_links}{to refresh the database queried list of links between experiments and devices (either from the device or experiment perspective)}
}
\description{
Device Info Server
}
