% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{ll_get_device_data_logs}
\alias{ll_get_device_data_logs}
\title{Retrieve data logs}
\usage{
ll_get_device_data_logs(
  group_id = default(group_id),
  filter = NULL,
  select = c(exp_id, device_id, device_name, device_data_log_id, datetime, data_idx,
    starts_with("data_key"), starts_with("data_")),
  max_rows = NULL,
  convert_to_TZ = Sys.timezone(),
  con = default(con),
  quiet = default(quiet)
)
}
\arguments{
\item{filter}{what filter conditions to apply, if any (forwarded to \link[dplyr]{filter})}

\item{select}{what columns to select (forwarded to \link[dplyr]{select}), by default a selection of the most commonly used columns}

\item{max_rows}{if provided, only selects the indicated number of rows (more efficient this way than part of the filter)}

\item{convert_to_TZ}{converts the log_datetime to the provided timezone (by default the system's local timezone stored in \code{Sys.timezone()}).}
}
\value{
device data logs
}
\description{
Returns data logs from the database joined with experiments, devices, and experiment_device_data (for prefix only) tables so filter conditions can be applied on these as well. Sorted by descending order (i.e. latest record first). Note that the group_id filter restricts the selected experiments, device data may theoretically come from devices not (or no longer) associated with the same group although this is rarely the case.
}
\seealso{
Other data logs functions: 
\code{\link{ll_plot_device_data_logs}()},
\code{\link{ll_read_device_data_logs_from_file}()},
\code{\link{ll_write_device_data_logs_to_file}()}
}
\concept{data logs functions}
