% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_cloud.R
\name{ll_get_devices_cloud_info}
\alias{ll_get_devices_cloud_info}
\title{Get general device information}
\usage{
ll_get_devices_cloud_info(devices = ll_get_devices(group_id = group_id, con =
  con), group_id = default(group_id), con = default(con),
  access_token = default(access_token), convert_to_TZ = Sys.getenv("TZ"),
  quiet = default(quiet))
}
\arguments{
\item{access_token}{the access token for the accout}

\item{particle_id}{the ID(s) of the particle device(s)}
}
\value{
nested data frame (converted from JSON)
}
\description{
Get information from the particle cloud about devices.
}
